+++ mirrord-layer/src/common.rs
@@ -1,4 +1,4 @@
-use std::{io::SeekFrom, os::unix::io::RawFd, path::PathBuf};
+use std::{io::SeekFrom, path::PathBuf};
 
 use mirrord_protocol::{
     CloseFileResponse, OpenFileResponse, OpenOptionsInternal, ReadFileResponse, SeekFileResponse,
@@ -38,7 +38,7 @@ pub struct OpenFileHook {
 
 #[derive(Debug)]
 pub struct OpenRelativeFileHook {
-    pub(crate) relative_fd: RawFd,
+    pub(crate) relative_fd: usize,
     pub(crate) path: PathBuf,
     pub(crate) file_channel_tx: oneshot::Sender<OpenFileResponse>,
     pub(crate) open_options: OpenOptionsInternal,
@@ -46,28 +46,28 @@ pub struct OpenRelativeFileHook {
 
 #[derive(Debug)]
 pub struct ReadFileHook {
-    pub(crate) fd: RawFd,
+    pub(crate) fd: usize,
     pub(crate) buffer_size: usize,
     pub(crate) file_channel_tx: oneshot::Sender<ReadFileResponse>,
 }
 
 #[derive(Debug)]
 pub struct SeekFileHook {
-    pub(crate) fd: RawFd,
+    pub(crate) fd: usize,
     pub(crate) seek_from: SeekFrom,
     pub(crate) file_channel_tx: oneshot::Sender<SeekFileResponse>,
 }
 
 #[derive(Debug)]
 pub struct WriteFileHook {
-    pub(crate) fd: RawFd,
+    pub(crate) fd: usize,
     pub(crate) write_bytes: Vec<u8>,
     pub(crate) file_channel_tx: oneshot::Sender<WriteFileResponse>,
 }
 
 #[derive(Debug)]
 pub struct CloseFileHook {
-    pub(crate) fd: RawFd,
+    pub(crate) fd: usize,
     pub(crate) file_channel_tx: oneshot::Sender<CloseFileResponse>,
 }